+++ hkdf/hkdf.ml
@@ -25,7 +25,7 @@ module Make (H : Digestif.S) : S = struct
       let nc = String.make 1 (Char.unsafe_chr n) in
       H.(to_raw_string (hmac_string ~key:prk (String.concat "" [last ; info ; nc])))
     in
-    let n = succ (len / H.digest_size) in
+    let n = (len + H.digest_size - 1) / H.digest_size in
     let rec compute acc count = match count, acc with
       | c, xs when c > n -> String.concat "" (List.rev xs)
       | c, x::_ -> compute (t c x :: acc) (succ c)